backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1
biopython==1.81
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2